Consistently handle the 'disabled' attr in inputs
As discussed in https://community.zammad.org/t/automating-creation-of-custom-object-attributes/3831/5

This commit ensures that the generic input widgets handle the `disabled`
attribute consistently. Before only the checkbox, select and input widgets respected
this attribute.

I've added it to the following widgets

* tree/searchable select
* textarea
* datetime
* date

Now it is possible to define a screens policy like so:
```
screens: {
"edit"=>{
"ticket.customer"=>{"shown"=>true, "required"=>false},
"ticket.agent"=>{
  "shown"=>true,
  "required"=>false,
  "disabled"=>true # <--- add this
}}}
```

Then the fields will show up on the Agent's sidebar with the customer
submitted values, but the Agent will not be able to change the values.

This is not the same as real permissions, as presumably a manually
crafted API request could circumvent the disabled attribute.

But this is a first step towards that direction, and very useful for us
in prod right now.